hedge_requirements.txt
# Runtime basics
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
tenacity==9.1.2
pydantic==2.11.10

# Async/network
aiohttp==3.12.15
websocket-client==1.8.0
websockets==15.0.1

# Crypto
pycryptodome==3.23.0
ecdsa==0.18.0
cryptography==46.0.2

# Paradex SDK deps (pin to avoid conflicts)
paradex_py==0.4.9
starknet-py==0.22.0
web3==6.20.3
eth-account==0.11.3
hexbytes==0.3.1
ckzg==1.0.2

# NOTE:
# - Lighter SDK is installed separately to avoid dependency resolver
#   upgrading eth-account to >=0.13 and breaking paradex_py/web3.
#   After installing the above, install lighter with:
#   pip install --no-deps git+https://github.com/elliottech/lighter-python.git@d0009799970aad54ebb940aa3dc90cbc00028c54


